requirements.txt
python-dotenv~=0.19.2
discord~=1.7.3
psutil~=5.9.0
pincer<=0.15.3
